requirements.txt
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
python-dotenv>=1.0.0
requests>=2.31.0
pytz>=2023.3

protobuf~=6.31.1
google-api-python-client